fix: Use float rather than int for node priority+ remove query limit

+++ src/schemas/node_view.py
@@ -15,7 +15,7 @@ class NodeView(BaseModel):
     parent_id: Optional[int] = None
     type: int
     content: Optional[NodeContent] = None
-    priority: int
+    priority: float
     due: int
     deleted_at: Optional[int] = None
     data: Optional[NodeData] = None
